+++ src/redox.rs
@@ -1,24 +1,24 @@
 use core::ffi::c_void;
 use core::ptr;
 
-/// Raw window handle for Redox OS.
+/// Raw window handle for the Redox operating system.
 ///
 /// ## Construction
 /// ```
-/// # use raw_window_handle::redox::RedoxHandle;
-/// let mut handle = RedoxHandle::empty();
-///  /* set fields */
+/// # use raw_window_handle::OrbitalHandle;
+/// let mut handle = OrbitalHandle::empty();
+/// /* set fields */
 /// ```
 #[non_exhaustive]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct RedoxHandle {
+pub struct OrbitalHandle {
     /// A pointer to an orbclient window.
     pub window: *mut c_void,
 }
 
-impl RedoxHandle {
-    pub fn empty() -> RedoxHandle {
-        RedoxHandle {
+impl OrbitalHandle {
+    pub fn empty() -> OrbitalHandle {
+        OrbitalHandle {
             window: ptr::null_mut(),
         }
     }